requirements.txt
requests==2.32.5
pandas==3.0.1
SQLAlchemy==2.0.48
PyYAML==6.0.3
dotenv==0.9.9
psycopg2-binary==2.9.11